+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 * **confxxl:** Scheduling API
 * **confxxl:** add xxl-job local Scheduling API
 
+### Fix
+
+* **confxxl:** xxl-job response
+
 
 <a name="v0.3.4"></a>
 ## [v0.3.4](https://github.com/xoctopus/confx/compare/v0.3.3...v0.3.4)